adding edit-score command

+++ commands/reset-scores.js
@@ -3,28 +3,36 @@ var postScore = require('./post-score');
 
 module.exports = {
   slash: true,
-  // testOnly: true,
-  testOnly: false,
+  testOnly: true,
+  // testOnly: false,
   guildOnly: true,
   description: 'Reset/clear scores and teams for Competition Corner.  This can be used to start a new week.',
-  callback: async ({client}) => {
-    const db = new JSONdb('db.json');
-    const archive = new JSONdb('archive.json');
+  callback: async ({client, channel}) => {
+    let retVal;
     
-    archive.storage.push(db.storage);
-    archive.sync();
+    if(channel.name !== 'competition-corner') {
+      retVal = 'The reset-scores slash command can only be used in the competition-corner channel.';
+    } else {
+      const db = new JSONdb('db.json');
+      const archive = new JSONdb('archive.json');
+      
+      archive.storage.push(db.storage);
+      archive.sync();
 
-    // get scores from db
-    const scores = db.get('scores') ? JSON.parse(db.get('scores')) : [];
+      // get scores from db
+      const scores = db.get('scores') ? JSON.parse(db.get('scores')) : [];
 
-    // clear scores
-    db.delete('scores');
-    // clear teams
-    db.delete('teams');
+      // clear scores
+      db.delete('scores');
+      // clear teams
+      db.delete('teams');
 
-    //post to competition channel pinned message
-    await postScore.editChannel([], client);
+      //post to competition channel pinned message
+      await postScore.editChannel([], client);
 
-    return "Scores and Teams have been reset."
+      retVal = "Scores and Teams have been reset."
+    }
+
+    return retVal;
   },
 }
